fix(TargetESP): unbox NumberSetting.get() before int cast
NumberSetting extends Setting<Double>, so .get() returns a boxed
Double. Casting a Double directly to int doesn't compile -
javac requires unboxing to a primitive double first.

+++ src/main/java/dev/blumdlc/client/modules/impl/TargetESP.java
@@ -103,7 +103,7 @@ public final class TargetESP extends Module {
 			.translate(-px, -py, 0.0f);
 
 		// 4. Draw the quad through the existing texture renderer.
-		int alpha = clampAlpha((int) brightness.get());
+		int alpha = clampAlpha((int) brightness.get().doubleValue());
 		QuadColorState corners = paletteFor(color.get(), angle, alpha);
 
 		Builder.texture()
